Add owner check to poll ending

+++ cogs/Polls.py
@@ -2,7 +2,7 @@ import discord
 from discord import Interaction, app_commands, ui
 from discord.ext import commands
 from WorstBot import WorstBot
-from modules import EmbedGen, Graphs, Paginators
+from modules import EmbedGen, Graphs, Paginators, Converters
 
 class Responses(ui.Select):
     def __init__(self, vote_id: int, options: list[discord.SelectOption], accepted_responses: int):
@@ -171,6 +171,9 @@ class Poll(commands.GroupCog, name = "poll"):
 
     @app_commands.command(name = "end", description = "End a poll and show the results")
     async def VoteEnd(self, interaction: Interaction, poll_id: int):
+        owner = await self.bot.fetchval("SELECT author FROM votes WHERE vote_id=$1", poll_id)
+        if interaction.user.id != owner:
+            return await interaction.response.send_message(f"This is not your poll to end, please use {Converters.to_command_mention(self.VoteResults, interaction.guild)}")
         await self.results(interaction, poll_id, ephemeral = False)
         channel, message = await self.bot.fetchrow("DELETE FROM votes WHERE vote_id = $1 RETURNING channel, message_id", poll_id)  # type: int, int
         channel: discord.PartialMessageable = self.bot.get_partial_messageable(channel)
